requirements.txt
# Core Framework
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# API Clients
openai>=1.0.0
google-genai>=0.3.0
requests>=2.31.0

# Web Framework
flask>=3.0.0

# Image Processing
Pillow>=10.0.0

# Evolution API (WhatsApp)
evolutionapi>=0.0.8

# Environment & Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
